Sunday Afternoon Push
Best update yet.
So the timer is now working, and resume/ pauses with the pause button

Added some basic colour changes of the background just to help simulate the idea.

Added input boxes to feed work and rest times into the timer. Just need to plug em into the back.

I seem to have an issue with the code changing the colour scheme before the timer kicks off, but I think I need to step away and return to fix that nicely.