Sunday Morning Push
I've got the work and rest time inputs on screen to implement in the pomodoro timer.

If you extend or diminish remaining time whilst the timer is running, this should up date instantly now.

Could of routes in the logic to test out. Would like to implement a switch to  have custom volume levels between work and rest mode.